Modify attribute id integer to long

+++ src/main/java/cat/udl/eps/entsoftarch/webingogeiapi/repository/GameRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 
 @RepositoryRestResource
-public interface GameRepository extends PagingAndSortingRepository<Game, Integer> {
+public interface GameRepository extends PagingAndSortingRepository<Game, Long> {
 
 }
